# [TESTERS NEEDED] SPU: Reimplement SPURS limiter