* Release thread semaphore upon successful SPURS Idle state setting, wait for semaphore before supposed SPURS task acquire.