@Augusto7743 2024-10-20
All good with you elad335 ? I wish all good in your life. Thanks for doing that good act.

That PR does any optimization when SPUs are extremely busy using each cpu core in 100 % ?
I see few to much speed optimization in some games.
Pro Evolution Soccer 2018 before your lastest spus optimizations code not was possible run in good performance in an cpu amd fx 6 cores running between 15 to 25 FPS. Now the game run between 35 to 46 FPS being playable !
I see moderate performance gain in Motorstorm Pacific Rift.
However has issues in few games as if SPU limiter not allow use the cpu in 100 %. Thus for any game not running 100 % need maximum spurs to above 4 or default value unlimited.

GT6 need an more recent amd zen cpu, but testing in an amd fx 6 cores if configuring the maximum spurs to less of 3 now is possible boot the game and the cpu usage drop to less of 75 % so having less performance. Need maximum spurs above of 4 to the cpu to be used 100 %. Perhaps that line information was correct to post.