@Augusto7743 2024-10-22
@elad335
Your recent optimizations code SPURS limiter is awesome.
Previous builds without SPURS limiter only is possible boot some games if SPURs is value unlimited or using other values happen crash.
Now is possible boot several games using SPURs with low value as 2 or 3 without crash.

I see here when testing some games three behaviors in current builds using SPURS limiter code :

- Medium to high performance optimization without crashes and using less cpu selecting SPURs to 2 or 3. Pro Evolution Soccer 2018 has high performance optimization.
- Some games run with less or same performance not using totally the cpu as was in previous builds before was added the SPURs limiter code even if SPURs value is unlimited. For some games using SPURs unlimited restore the 100 % gameplay speed and for others games the performance is below of 100 % not using totally the cpu.
- Others games nothing happen. Need for Speed Hot Pursuit is an example. Run in same performance without crashes without any optimization as if the SPURs limiter code not work.

Is as if the SPU limiter code need to be an option in the RPCS3 GUI Advanced menu to choice if enable for some games and disable for other games so avoiding errors.