@Augusto7743 2024-10-22
I remember the FPS performance in some games using previous build and with current build using your SPURs limiter code totally sure are more faster for some games.
Here the CPU is an amd fx 6300 6 cores and even so has performance optimization at point games unplayable are now playable being the detail important to post about it.

Pro Evolution Soccer 2018 never had an performance above 24 FPS not possible to play. Now the game run between 35 to 46 FPS being playable in my cpu.
MOTORSTORM PACIFIC RIFT NPUA70033 Demo 2 never was possible an good speed to play and was performance below of MOTORSTORM PACIFIC RIFT NPUA70022 Demo 1. Both are much slow in my cpu and never was above 12~14 FPS in previous builds. Now only for test both are between 14 to 18 FPS at point is possible play in medium speed not being exactly slow as was before.

AMD FX not run Need for Speed Hot Pursuit and GT6, but for test Need for Speed Hot Pursuit continue being the same performance 12 FPS as if the SPURs limiter code not does anything even using SPURs 2 or 3 and GT6 run more slow below 10 FPS with cpu usage never being above 75 % and before in previous builds run in 14 FPS cpu usage between 90 and 100 %.
Is good an user using current cpus having Need for Speed Hot Pursuit and GT6 post about the performance using current build.

I understand AMD FX not is really an good cpu for all games, but now run all games in good speed less NFS Hot Pursuit and GT6 both need an better cpu.
Thus being if your code does an good optimization for AMD FX I had to post about it.

You need logs ?